Moved enums definitions from main.cpp to enums.h. Moved code related to these enums to common.h.
